+++ internal/models/transaction.go
@@ -6,9 +6,9 @@ import (
 
 // UploadTransaction represents a file upload transaction.
 type UploadTransaction struct {
-	ID        int       `db:"id"`
+	ID        int64     `db:"id"`
 	UserID    int       `db:"user_id"`
-	FolderID  int       `db:"folder_id"`
+	FolderID  int64     `db:"folder_id"`
 	TotalSize int64     `db:"total_size"`
 	Status    string    `db:"status"`
 	CreatedAt time.Time `db:"created_at"`
